torch nightly: disambiguate Tensor::argsort overload on LibTorch 2.4
LibTorch 2.4 introduced a second argsort overload with a leading bool
stable parameter. Calling `noise.argsort(/*dim=*/1)` with an int literal
is ambiguous between:

  argsort(int64_t dim, bool descending)
  argsort(bool stable, int64_t dim, bool descending)

Cast the dim argument to int64_t to pick the original overload
explicitly. 2.5.x has the same dual overloads so this fixes both
nightly matrix entries.

+++ src/fsgea_phenotype_gpu.hpp
@@ -67,7 +67,9 @@ inline torch::Tensor permuteLabels(
     auto gen = at::detail::createCPUGenerator(static_cast<std::uint64_t>(seed));
     auto noise = torch::rand({B, S}, gen,
         torch::TensorOptions().dtype(torch::kFloat32)).to(device);
-    auto const perm_idx = noise.argsort(/*dim=*/1);   // [B, S], int64
+    // Explicit int64_t disambiguates argsort(int64_t, bool) from
+    // argsort(bool, int64_t, bool) on LibTorch 2.4.x.
+    auto const perm_idx = noise.argsort(static_cast<std::int64_t>(1));  // [B, S], int64
     // labels.unsqueeze(0).expand({B, S}) is shape-compatible; gather picks
     // the permuted labels per row.
     return labels.unsqueeze(0).expand({B, S}).gather(/*dim=*/1, perm_idx);
